games_requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
bcrypt==3.2.0
bidict==0.21.2
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
discord==1.0.1
discord.py==1.5.1
dnspython==1.16.0
Flask==1.1.2
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.4.4
gevent==20.12.1
gevent-websocket==0.10.1
greenlet==0.4.17
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.1
MarkupSafe==1.1.1
multidict==4.7.6
multipledispatch==0.6.0
numpy==1.20.3
pandas==1.2.4
Pillow==8.0.1
pycparser==2.20
pymongo==3.12.0
PySide2==5.15.2
python-dateutil==2.8.2
python-engineio==4.0.0
python-socketio==5.0.4
pythonWordArt==0.9
pytz==2021.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.7.0
selenium==3.141.0
shiboken2==5.15.2
six==1.15.0
sklearn==0.0
spotipy==2.19.0
SQLAlchemy==1.3.22
threadpoolctl==2.2.0
torch==1.8.1
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
yarl==1.5.1
youtube-dl==2021.6.6
youtube-python==1.0.13
zope.event==4.5.0
zope.interface==5.2.0